requirements.txt
fastapi
httpx
uvicorn
python-dotenv
jinja2